feat: refactor ImageUploader to use custom hook for improved file handling and upload logic
